requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.1
PyJWT~=2.7.0
python-dotenv~=1.0.0
google-api-python-client~=2.111.0
django-cors-headers~=4.3.1
Pillow~=10.2.0